+++ CHANGELOG.md
@@ -10,6 +10,10 @@ when a pull request with a release label merges into `main`.
 
 ## [Unreleased]
 
+### Added
+
+- Allow an unrestricted root session to revoke another user's session using the built-in root capability set.
+
 ## [v0.8.0] - 2026-09-20
 
 ### Added